[241001] refactor: graphql 코드 분리
